[filesystems] Move FileSystemProvider meta-class in __init__
